Report from Robby Dermody:
Thanks for making py-asterisk. I'm upgrading my systems to Asterisk 10.3.0 (the
newest stable) and notice that BaseManager doesn't work with it:

_AST_BANNERS = [
    'Asterisk Call Manager/1.0\r\n'
    'Asterisk Call Manager/1.1\r\n'
]

'Asterisk Call Manager/1.2' needs to be in that list too. Not sure if there are
any other changes due to the new release, but it would be awesome if you could
issue a new release with this fixed in it. Double-awesome if it was on github.
;)