Patch from Alexis de Lattre to allow duplicate header names.
But I had a difficulty using py-asterisk to generate the Originate request that
I need ; here is an example of the request I need :

Action: Originate
Channel: sip/7ltyb7
Exten: 0699089245
Context: default
Priority: 1
Timeout: 3600000
Variable: XIVO_ORIGAPPLI=OrigDial
Variable: XIVO_ORIG_CID_NAME=Alexis-
Variable: XIVO_ORIG_CID_NUM=21
CallerID: "0699089245"<0699089245>

The problem is that, with the current code of my-asterisk, I can't have several
"Variable" lines in my Originate request. So I wrote this patch, to be able to
use a list for the "variable" argument of the Originate function (and not just
use a string) ; I tested it and it works well. But, as you can see in my patch
enclosed, the modification is not in the code of the Originate() fonction, it's
in the code of the _write_action() fonction, which is used my many other
functions.