Fix for extra newlines in some SMS messages
Some SMS messages have an extra '\r\n' line.  This extra line kills the connection to the Asterisk Manager because _read_packet() raises an exception if it receives a blank line.  

This fix checks the connection first when receiving an empty line before killing it.  If Ping() fails, it raises an exception which then triggers the exception that the Asterisk Manager has gone away.